requirements.txt
apache-airflow==2.10.2
pandas==1.3.5
openpyxl==3.0.10
psycopg2-binary==2.9.9
scikit-learn==1.0.2
sqlalchemy==1.4.46
